fix: analytics
