+++ client/src/pages/_document.tsx
@@ -1,8 +1,5 @@
 import * as React from 'react';
 import Document, { Head, Html, Main, NextScript } from 'next/document';
-import Script from 'next/script';
-
-const enableAnalytics = process.env.NODE_ENV === 'production';
 
 export default class extends Document {
   render() {
@@ -10,16 +7,6 @@ export default class extends Document {
       <Html lang="en">
         <Head>
           <link rel="shortcut icon" href="https://rs.school/favicon.ico" />
-
-          {enableAnalytics && <Script async src="https://www.googletagmanager.com/gtag/js?id=G-WJLHZ9CCXJ" />}
-          {enableAnalytics && <script dangerouslySetInnerHTML={{ __html: gaJsCode }} />}
-          {enableAnalytics && (
-            <script
-              defer
-              src="https://static.cloudflareinsights.com/beacon.min.js"
-              data-cf-beacon={'{"token": "e607238d732c4713b01b851ed3df61c2"}'}
-            />
-          )}
         </Head>
         <body>
           <Main />
@@ -29,11 +16,3 @@ export default class extends Document {
     );
   }
 }
-
-const gaJsCode = `
-  window.dataLayer = window.dataLayer || [];
-  function gtag(){dataLayer.push(arguments);}
-  gtag('js', new Date());
-
-  gtag('config', 'G-WJLHZ9CCXJ');
-`;
